10 /18 기능수정 1. Tower, FriendlyUnit의 모든 업데이트는 TowerController에서 Unitask로 딜레이를 주어 실행시킴 => https://blog.unity.com/kr/engine-platform/10000-update-calls 참조 2. 1번과 같은 이유로 EnemyUnit역시 WaveManager에서 스폰되기 때문에 이 매니저가 Enemy들을 Update시키게 된다. 3. Tower, Unit 누를 때 UI를 OnPointerUp에서 해줌 => 원래는 투명 UI가 확인해서 UI를 켜주었음 이럴 필요가 없어짐

+++ TowerDefense/Assets/Scripts/ProjectileControl/Projectile.cs
@@ -13,7 +13,7 @@ namespace ProjectileControl
         private ParticleSystem _hitParticle;
         private Rigidbody _rigid;
 
-        private ushort _damage;
+        private int _damage;
         private float _gravity;
         private Vector3 _curPos;
         private Vector3 _startPos;
@@ -90,7 +90,7 @@ namespace ProjectileControl
             _rigid.MoveRotation(Quaternion.LookRotation(dir));
         }
 
-        public virtual void Init(ushort dmg, Transform t)
+        public virtual void Init(int dmg, Transform t)
         {
             _damage = dmg;
             target = t;
